refactor: move policy process to process.py
